requirements.txt
tensorflow>=2.10
flask
pillow
numpy
